web/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-environ==0.4.5
django-js-asset==1.2.2
django-mptt==0.12.0
gunicorn==20.0.4
idna==2.10
psycopg2==2.8.6
pyTelegramBotAPI==3.7.6
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
vk-api==11.9.1
